feat: support producing documents
