+++ producer/Cargo.toml
@@ -13,6 +13,9 @@ did_key = { path = "../did_key" }
 did_web = { path = "../did_web" }
 shared = { path = "../shared" }
 
+oid4vc-core = { git = "https://github.com/impierce/openid4vc.git", rev = "ebd7a83" }
+
+anyhow = "1.0"
 hex = "0.4"
 identity_iota.workspace = true
 identity_stronghold.workspace = true
@@ -20,6 +23,9 @@ iota-sdk.workspace = true
 iota_stronghold = { version = "2.0" }
 log = "0.4"
 rand = "0.8"
+serde = "1.0"
+serde_json = "1.0"
+tokio.workspace = true
 url = "2.5"
 
 [dev-dependencies]
