+++ producer/src/lib.rs
@@ -1,111 +1,5 @@
+pub mod did_document;
 pub mod secret_manager;
 pub mod signature;
 
-use crate::secret_manager::SecretManager;
-
-use identity_iota::{document::CoreDocument, storage::KeyId};
-use shared::JwkStorageWrapper;
-
-pub enum Method {
-    Jwk,
-    Key,
-    Web,
-}
-
-pub async fn get_document(
-    secret_manager: SecretManager,
-    method: Method,
-    host: Option<url::Host>, // only used for `did:web` --> find a better way to provide this
-    port: Option<u16>,       // only used for `did:web` --> find a better way to provide this
-) -> std::result::Result<CoreDocument, std::io::Error> {
-    let storage = JwkStorageWrapper::Stronghold(secret_manager.stronghold_storage);
-
-    // TODO: remove this hard-coded value and replace it with: get_key_id(method_digest)
-    let key_id = KeyId::new("9O66nzWqYYy1LmmiOudOlh2SMIaUWoTS");
-
-    let core_document: Option<CoreDocument> = match method {
-        Method::Jwk => {
-            let core_document = did_jwk::producer::produce_did_jwk(storage, key_id.as_str())
-                .await
-                .unwrap();
-            Some(core_document)
-        }
-        Method::Key => {
-            let core_document = did_key::producer::produce_did_from_key(storage, &key_id).await.unwrap();
-            Some(core_document)
-        }
-        Method::Web => {
-            let core_document =
-                did_web::producer::produce_did_web(storage, &key_id, host.expect("host not specified"), port)
-                    .await
-                    .unwrap();
-            Some(core_document)
-        }
-    };
-
-    match core_document {
-        Some(core_document) => Ok(core_document),
-        None => Err(std::io::Error::other("No core_document produced")),
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    use identity_iota::core::{json, ToJson};
-    use shared::test_utils::random_stronghold_path;
-
-    #[tokio::test]
-    #[ignore]
-    async fn create_document_from_generated_stronghold() {
-        let secret_manager = SecretManager::generate(
-            random_stronghold_path().to_str().unwrap().to_string(),
-            "secure_password".to_string(),
-        );
-
-        let document = get_document(
-            secret_manager.unwrap(),
-            Method::Web,
-            Some(url::Host::parse("localhost").unwrap()),
-            Some(8080),
-        )
-        .await;
-
-        println!("document: {}", document.as_ref().unwrap().to_json_pretty().unwrap());
-        assert!(document.is_ok())
-    }
-
-    #[tokio::test]
-    async fn recreate_expected_document_from_existing_stronghold() {
-        let secret_manager =
-            SecretManager::load("tests/res/test.stronghold".to_string(), "secure_password".to_string());
-
-        let document = get_document(
-            secret_manager.unwrap(),
-            Method::Web,
-            Some(url::Host::parse("localhost").unwrap()),
-            Some(8080),
-        )
-        .await;
-
-        assert_eq!(
-            document
-                .unwrap()
-                .verification_method()
-                .first()
-                .unwrap()
-                .to_json_value()
-                .unwrap()
-                .get("publicKeyJwk")
-                .unwrap(),
-            &json!({
-                "kty": "OKP",
-                "alg": "EdDSA",
-                "kid": "aHq-0PIf6_ljLhyx4W86Gviqb-671OAI67E6vXpZc7Q",
-                "crv": "Ed25519",
-                "x": "P2BkYS6z4UHmsxn6FX1oHsyx7eiUSFEMJ1D_RC8M0-w"
-            })
-        )
-    }
-}
+pub use crate::secret_manager::SecretManager;
