+++ src/lib.rs
@@ -1,0 +1,2 @@
+// pub use producer;
+pub use producer::SecretManager;
